refactoring the ticketing system into booking, model and woker packages
