+++ main.go
@@ -3,173 +3,14 @@ package main
 import (
 	"context"
 	"fmt"
+	"golang-bootcamp-July/booking"
+	"golang-bootcamp-July/worker"
 	"log"
-	"math/rand"
 	"strings"
 	"sync"
 	"sync/atomic"
 	"time"
 )
-
-type TicketSystem struct {
-	totalTickets     int32
-	availableTickets int32
-	mu               sync.Mutex
-	bookings         map[string]*Booking
-}
-
-type Booking struct {
-	ID        string
-	UserID    string
-	Timestamp time.Time
-}
-
-type BookingResult struct {
-	UserID    string
-	Success   bool
-	BookingID string
-	Error     error
-}
-
-func NewTicketSystem(totalTickets int32) *TicketSystem {
-	return &TicketSystem{
-		totalTickets:     totalTickets,
-		availableTickets: totalTickets,
-		bookings:         make(map[string]*Booking),
-	}
-}
-
-func (ts *TicketSystem) BookTicket(userID string) BookingResult {
-	// network latency
-	time.Sleep(time.Microsecond * time.Duration(rand.Intn(100)))
-
-	// non-blocking read
-	if atomic.LoadInt32(&ts.availableTickets) <= 0 {
-		return BookingResult{
-			UserID:  userID,
-			Success: false,
-			Error:   fmt.Errorf("no tickets available"),
-		}
-	}
-
-	// locking
-	ts.mu.Lock()
-	defer ts.mu.Unlock()
-
-	// double check
-	if ts.availableTickets <= 0 {
-		return BookingResult{
-			UserID:  userID,
-			Success: false,
-			Error:   fmt.Errorf("no tickets available"),
-		}
-	}
-
-	// create
-	bookingID := fmt.Sprintf("BOOK-%s-%d", userID, time.Now().UnixNano())
-	booking := &Booking{
-		ID:        bookingID,
-		UserID:    userID,
-		Timestamp: time.Now(),
-	}
-
-	// update
-	ts.bookings[bookingID] = booking
-	atomic.AddInt32(&ts.availableTickets, -1)
-
-	return BookingResult{
-		UserID:    userID,
-		Success:   true,
-		BookingID: bookingID,
-	}
-}
-
-func (ts *TicketSystem) GetStats() (totalBooked int32, available int32) {
-	available = atomic.LoadInt32(&ts.availableTickets)
-	totalBooked = ts.totalTickets - available
-	return
-}
-
-// worker to process request from a channel
-type Worker struct {
-	id       int
-	requests <-chan string
-	results  chan<- BookingResult
-	ts       *TicketSystem
-	wg       *sync.WaitGroup
-}
-
-// start booking requests
-func (w *Worker) Start(ctx context.Context) {
-	defer w.wg.Done()
-
-	for {
-		select {
-		case userID, ok := <-w.requests:
-			if !ok {
-				// channel closed, worker should exit
-				return
-			}
-
-			result := w.ts.BookTicket(userID)
-
-			// send result
-			select {
-			case w.results <- result:
-			case <-ctx.Done():
-				return
-			}
-
-		case <-ctx.Done():
-			return
-		}
-	}
-}
-
-type WorkerPool struct {
-	workers  []*Worker
-	requests chan string
-	results  chan BookingResult
-	wg       sync.WaitGroup
-}
-
-func NewWorkerPool(numWorkers int, ts *TicketSystem) *WorkerPool {
-	pool := &WorkerPool{
-		requests: make(chan string, numWorkers*2),
-		results:  make(chan BookingResult, numWorkers*10),
-		workers:  make([]*Worker, numWorkers),
-	}
-
-	for i := 0; i < numWorkers; i++ {
-		pool.workers[i] = &Worker{
-			id:       i,
-			requests: pool.requests,
-			results:  pool.results,
-			ts:       ts,
-			wg:       &pool.wg,
-		}
-	}
-
-	return pool
-}
-
-// start all the workers
-func (p *WorkerPool) Start(ctx context.Context) {
-	for _, worker := range p.workers {
-		p.wg.Add(1)
-		go worker.Start(ctx)
-	}
-}
-
-func (p *WorkerPool) Stop() {
-	close(p.requests)
-	p.wg.Wait()
-	close(p.results)
-}
-
-func (p *WorkerPool) SubmitRequest(userID string) {
-	p.requests <- userID
-}
 
 func main() {
 	const (
@@ -186,13 +27,13 @@ func main() {
 
 	startTime := time.Now()
 
-	ticketSystem := NewTicketSystem(totalTickets)
+	ticketSystem := booking.NewTicketSystem(totalTickets)
 
 	ctx, cancel := context.WithTimeout(context.Background(), 30*time.Second)
 	defer cancel()
 
-	workerPool := NewWorkerPool(numWorkers, ticketSystem)
-	workerPool.Start(ctx)
+	workerPool := worker.NewWorkerPool(numWorkers, ticketSystem)
+	workerPool.StartWorkerPool(ctx)
 
 	var (
 		successfulBookings int32
@@ -202,7 +43,7 @@ func main() {
 
 	resultsDone := make(chan bool)
 	go func() {
-		for result := range workerPool.results {
+		for result := range workerPool.Results() {
 			atomic.AddInt32(&processedRequests, 1)
 
 			if result.Success {
