+++ model/types.go
@@ -1,0 +1,18 @@
+package model
+
+import (
+	"time"
+)
+
+type Booking struct {
+	ID        string
+	UserID    string
+	Timestamp time.Time
+}
+
+type BookingResult struct {
+	UserID    string
+	Success   bool
+	BookingID string
+	Error     error
+}
